% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_in_europe}
\alias{covid_in_europe}
\title{covid_in_europe data set}
\format{
A data frame with 5 columns and 163 rows.
}
\source{
www.ecdc.europa.eu
}
\usage{
covid_in_europe
}
\description{
A data frame with with daily and cumulative cases of Covid infections and deaths in Europe since March 2021.
}
\keyword{datasets}
